refactor(mentor): 메인페이지의 Empty 관련 컴포넌트 위치 수정
